requirements.txt
dnspython==2.6.1
motor==3.4.0
pymongo==4.7.3
python-dotenv==1.0.1
